// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  MESA_DE_PARTES: 'MESA_DE_PARTES',
  STAFF_OFICINA: 'STAFF_OFICINA',
  JEFE_OFICINA: 'JEFE_OFICINA',
  GERENTE: 'GERENTE',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const DocumentStatus = {
  creado: 'creado',
  recibido: 'recibido',
  derivado: 'derivado',
  en_revision: 'en_revision',
  observado: 'observado',
  atendido: 'atendido',
  archivado: 'archivado',
  rechazado: 'rechazado'
} as const

export type DocumentStatus = (typeof DocumentStatus)[keyof typeof DocumentStatus]


export const ApplicantType = {
  natural: 'natural',
  juridica: 'juridica'
} as const

export type ApplicantType = (typeof ApplicantType)[keyof typeof ApplicantType]


export const DocumentType = {
  CARTA: 'CARTA',
  DIRECTIVA: 'DIRECTIVA',
  EXPOSICI: 'EXPOSICI',
  N_DE_MOTIVOS: 'N_DE_MOTIVOS',
  INFORME: 'INFORME',
  MEMORANDO: 'MEMORANDO',
  MEMORANDO_MULTIPLE: 'MEMORANDO_MULTIPLE',
  NOTA_DE_ELEVACI: 'NOTA_DE_ELEVACI',
  OFICIO: 'OFICIO',
  OFICIO_CIRCULAR: 'OFICIO_CIRCULAR',
  OFICIO_MULTIPLE: 'OFICIO_MULTIPLE',
  AYUDA_MEMORIA: 'AYUDA_MEMORIA',
  OTROS: 'OTROS',
  PROVEIDO: 'PROVEIDO',
  N_VICEMINISTERIAL: 'N_VICEMINISTERIAL',
  RESUMEN_EJECUTIVO: 'RESUMEN_EJECUTIVO',
  SOBRE_CERRADO: 'SOBRE_CERRADO',
  SOLICITUD: 'SOLICITUD'
} as const

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType]


export const OfficeType = {
  ALCALDIA: 'ALCALDIA',
  GERENCIA_MUNICIPAL: 'GERENCIA_MUNICIPAL',
  ORGANO_STAFF: 'ORGANO_STAFF',
  GERENCIA_LINEA: 'GERENCIA_LINEA',
  OFICINA_GENERAL: 'OFICINA_GENERAL',
  UNIDAD: 'UNIDAD',
  COMITE_CONSULTIVO: 'COMITE_CONSULTIVO'
} as const

export type OfficeType = (typeof OfficeType)[keyof typeof OfficeType]